requirements.txt
bottle==0.12.19
mysql-connector==2.2.9
mysql-connector-python==8.0.28